requirements.txt
absl-py==2.4.0
adal==1.2.7
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
appdirs==1.4.4
argcomplete==3.6.3
async-timeout==5.0.1
asyncssh==2.22.0
atpublic==7.0.0
attrs==26.1.0
azure-common==1.1.28
azure-core==1.39.0
azure-graphrbac==0.61.2
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==14.0.0
azure-mgmt-core==1.6.0
azure-mgmt-keyvault==11.0.0
azure-mgmt-network==30.0.0
azure-mgmt-resource==24.0.0
azure-mgmt-storage==24.0.0
azureml-core==1.61.0.post3
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==5.0.0
billiard==4.2.4
blinker==1.9.0
boto3==1.42.91
botocore==1.42.91
cachetools==7.0.5
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
configobj==5.0.9
contextlib2==21.6.0
contourpy==1.3.2
cryptography==46.0.7
cycler==0.12.1
databricks-sdk==0.102.0
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==1.1.0
durationpy==0.10
dvc==3.67.1
dvc-data==3.18.3
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.3.1
fastapi==0.136.0
filelock==3.28.0
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
flatten-dict==0.4.2
flufl.lock==9.0.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-core==2.5.1
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.4.0
gto==1.9.0
gunicorn==25.3.0
h11==0.16.0
huey==2.6.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.3
importlib_metadata==8.7.1
isodate==0.7.2
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonpickle==4.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kagglehub==1.0.0
kagglesdk==0.1.21
kiwisolver==1.5.0
knack==0.12.0
kombu==5.6.2
kubernetes==35.0.0
lazy-loader==0.5
llvmlite==0.47.0
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.4
mlflow==3.11.1
mlflow-skinny==3.11.1
mlflow-tracing==3.11.1
mpmath==1.3.0
msal==1.36.0
msal-extensions==1.3.1
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.7.1
ndg-httpsclient==0.5.1
networkx==3.4.2
numba==0.65.0
numpy==1.26.4
oauthlib==3.3.1
omegaconf==2.3.0
opencv-python-headless==4.8.1.78
opentelemetry-api==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
opt_einsum==3.4.0
orjson==3.11.8
packaging==25.0
pandas==2.3.3
paramiko==3.5.1
pathspec==0.12.1
pillow==12.2.0
pkginfo==1.12.1.2
platformdirs==4.9.6
pooch==1.9.0
prettytable==3.17.0
prometheus_client==0.25.0
prometheus_flask_exporter==0.23.2
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.2
protobuf==3.20.3
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
pydot==4.0.1
pygit2==1.18.2
Pygments==2.20.0
pygtrie==2.5.0
PyJWT==2.12.1
PyMatting==1.1.15
PyNaCl==1.6.2
pyOpenSSL==25.3.0
pyparsing==3.3.2
pysftp==0.2.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
rembg==2.0.69
requests==2.33.1
requests-auth-aws-sigv4==0.7
requests-oauthlib==2.0.0
rich==15.0.0
rpds-py==0.30.0
ruamel.yaml==0.19.1
s3transfer==0.16.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
scmrepo==3.6.2
seaborn==0.13.2
SecretStorage==3.5.0
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
skops==0.13.0
smmap==5.0.3
sounddevice==0.5.5
SQLAlchemy==2.0.49
sqlparse==0.5.5
sqltrie==0.11.2
starlette==1.0.0
sympy==1.14.0
tabulate==0.10.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tomli==2.4.1
tomlkit==0.14.0
torch==2.2.2+cpu
torchvision==0.17.2+cpu
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.44.0
vine==5.1.0
voluptuous==0.16.0
wcwidth==0.6.0
websocket-client==1.9.0
Werkzeug==3.1.8
yarl==1.23.0
zc.lockfile==4.0
zipp==3.23.1
